# Conflicts:
#   plugins/mdm/js/MDM.js
#   plugins/mdm/php/class.pluginmdmmodule.php

+++ server/includes/core/class.webappsession.php
@@ -5,7 +5,7 @@ require_once(BASE_PATH . 'config.php');
 require_once(BASE_PATH . 'defaults.php');
 require_once(BASE_PATH . 'server/includes/util.php');
 require_once(BASE_PATH . 'server/includes/exceptions/class.JSONException.php');
- 
+
 /**
  * A class with helper functions for the PHP session
  * @singleton
@@ -15,11 +15,11 @@ class WebAppSession
 	// Holds the instance of this class
 	// Use public static method getInstance() to retrieve its value
 	private static $instance = null;
-	
+
 	// Set to true when the the session has timed out
 	// Use public method hasTimedOut() to retrieve its value
 	private $timeout = false;
-	 
+
 	/**
 	 * Constructor
 	 */
@@ -38,14 +38,14 @@ class WebAppSession
 		$path = ini_get('session.cookie_path');
 		$domain = ini_get('session.cookie_domain');
 		session_set_cookie_params($lifetime, $path, $domain, $secure, true);
-		
+
 		// Start the session so we can use it for timeout checking
 		$this->start();
-		
+
 		if ( basename($_SERVER['PHP_SELF']) != 'grommunio.php' ){
 			//We will only check for timeout in the grommunio.php page
 			$this->setStartTime();
-		}else{
+		} else {
 			$this->checkForTimeout();
 		}
 	}
@@ -55,7 +55,7 @@ class WebAppSession
 	 * To force this class to be used as singleton, the constructor is private
 	 * and the single instance can be created/retrieved with this static method
 	 * e.g.: $session = WebAppSession::createInstance()
-	 * 
+	 *
 	 * @return WebAppSession the only available instance of this class
 	 */
 	public static function getInstance()
@@ -65,7 +65,7 @@ class WebAppSession
 		}
 		return WebAppSession::$instance;
 	}
-	
+
 	/**
 	 * The method getInstance() was first called createInstance(), so
 	 * for backward compatibility we also define createInstance as an alias
@@ -74,7 +74,7 @@ class WebAppSession
 	{
 		return WebAppSession::getInstance();
 	}
-	
+
 	/**
 	 * Starts the session
 	 */
@@ -82,7 +82,7 @@ class WebAppSession
 	{
 		session_start();
 	}
-	
+
 	/**
 	 * Closes the session
 	 */
@@ -90,7 +90,7 @@ class WebAppSession
 	 {
 	 	session_write_close();
 	 }
-	 
+
 	/**
 	 * Destroy the session
 	 */
@@ -100,7 +100,7 @@ class WebAppSession
 		if ( session_status() !== PHP_SESSION_ACTIVE ){
 			session_start();
 		}
-		
+
 		// Destroy the session cookie
 		if (isset($_COOKIE[session_name()])) {
 			setcookie(session_name(), '', time()-42000, '/');
@@ -110,32 +110,32 @@ class WebAppSession
 		$_SESSION = array();
 		session_destroy();
 	}
-	
+
 	/**
-	 * Sets the start time in the session to current timestamp. 
-	 * This is used to logout the user when CLIENT_TIMEOUT has 
+	 * Sets the start time in the session to current timestamp.
+	 * This is used to logout the user when CLIENT_TIMEOUT has
 	 * been set in config.php
 	 */
 	public function setStartTime()
 	{
 		$_SESSION['starttime'] = time();
 	}
-	
+
 	/**
-	 * gets the start time from the session. This is used to logout the user 
+	 * gets the start time from the session. This is used to logout the user
 	 * when CLIENT_TIMEOUT has been set in config.php
-	 * 
+	 *
 	 * @return integer|false The starttime (timestamp) when set, false otherwise
 	 */
 	public function getStartTime()
 	{
 		if ( isset($_SESSION['starttime']) ){
 			return $_SESSION['starttime'];
-		}else{
+		} else {
 			return false;
 		}
 	}
-	
+
 	/**
 	 * Checks if the session should timeout and destroys the session if it should.
 	 */
@@ -145,13 +145,17 @@ class WebAppSession
 			//Timeout was not set in configuration, so do nothing
 			return;
 		}
-		
+
+		if (array_key_exists('HTTP_CONTENT_TYPE', $_SERVER) && strpos($_SERVER['HTTP_CONTENT_TYPE'], 'application/json') === false) {
+			return;
+		}
+
 		$starttime = $this->getStartTime();
 		// let's add 5 seconds to the CLIENT_TIMEOUT to handle possible latency
 		if ( $starttime && (time()-$starttime > CLIENT_TIMEOUT+5) ){
 			$this->destroy();
 			$this->timeout = true;
-		}else{
+		} else {
 			try{
 				// decode JSON data
 				$requestJsonString = readData();
@@ -159,17 +163,18 @@ class WebAppSession
 			}catch(JSONException $e){
 				// Invalid json sent with the request
 				// Log the error and do nothing is best option
-				dump($e->getDisplayMessage());
+				dump($e->getMessage());
+				return;
 			}
 
 			$isReminderlistRequest = $this->isReminderListRequest($requestJson);
 			$isDestroySessionRequest = $this->isDestroySessionRequest($requestJson);
-			
+
 			if ( !$isReminderlistRequest && !$isDestroySessionRequest){
 				//Set a new session start time
 				$this->setStartTime();
-			}elseif ( $isDestroySessionRequest ){
-				// sessiondestroy is sent because of timeout at client side 
+			} elseif ( $isDestroySessionRequest ){
+				// sessiondestroy is sent because of timeout at client side
 				$this->timeout = true;
 				$this->destroy();
 			}
@@ -178,7 +183,7 @@ class WebAppSession
 
 	/**
 	 * Checks if the current request is a destroysession request
-	 * 
+	 *
 	 * @param array $requestJson The JSON that was sent as the current request
 	 * @return boolean
 	 */
@@ -201,7 +206,7 @@ class WebAppSession
 
 	/**
 	 * Checks if the current request is a reminderlist request
-	 * 
+	 *
 	 * @param array $requestJson The JSON that was sent as the current request
 	 * @return boolean
 	 */
@@ -209,10 +214,10 @@ class WebAppSession
 	{
 		return isset($requestJson) && isset($requestJson['zarafa']) && isset($requestJson['zarafa']['reminderlistmodule']);
 	}
-	
+
 	/**
 	 * Returns true if the current session has timed out, false otherwise
-	 * 
+	 *
 	 * @return boolean
 	 */
 	public function hasTimedOut()
